requirements.txt
pandas~=1.1.3
setuptools~=50.3.0
Cython~=0.29.21
pyyaml~=5.3.1